Fix mock quotes storage package
