+++ app/src/main/kotlin/io/github/dector/quotes/qoutes/QuotesModule.kt
@@ -8,7 +8,7 @@ import io.github.dector.quotes.R
 import io.github.dector.quotes.domain.Quote
 import io.github.dector.quotes.presentation.presenter.QuotesPresenter
 import io.github.dector.quotes.qoutes.presentation.view.QuotesView
-import io.github.dector.quotes.qoutes.storage.MockQuotesStorage
+import io.github.dector.qoutes.storage.MockQuotesStorage
 import io.github.dector.quotes.storage.IStorage
 import io.github.dector.quotes.usecases.IQuotesUseCase
 import io.github.dector.quotes.usecases.QuotesUseCase
